feat: Props modal
